+++ CHANGE-LOG.md
@@ -20,6 +20,11 @@ V2.19.0
 -   单元格选择功能，添加表头和列指示器效果 [详情]
 -   单元格选择功能，添加区域选择实例方法`setRangeCellSelection({ startRowKey,startColKey,endRowKey,endColKey,isScrollToStartCell })` [详情]
 -   右键菜单组件添加`hideContextmenu()`实例方法
+-   添加主体变量
+    -   @ve-table-head-indicator-active-text-color
+    -   @ve-table-body-indicator-active-text-color
+    -   @ve-table-header-cell-indicator-active-background-color
+    -   @ve-table-body-cell-indicator-active-background-color
 
 ### Bug Fixes